Develop (#5)
* Preparing for v1.0.4 development

* PER-237 updated ldap adapter search method
The new syntax to prepend values to the base was not being respected by the search method.  Found the bug and fixed it.  Also added a test case for using filters.  In addition removed a few comments from the files.

* PER-237 added logging statement to ldap adapter
Added a logging statement to record if the base has been updated.  This was done to assist in future troubleshooting efforts.

* PER-237 updated ldap adapter version number in pom

* PER-237 updated ldap adapter pom and changelog
Added to changelog the update to the ladp adapter.  Found an issue with the version number and correct.

* Added changelog and readme to ldap adapter

Co-authored-by: Benjamin Christenson <ben.christenson@kineticdata.com>